added resume parsing, gemini to parse resume data,  error handling to matches,applied page and added application close logic in ui

+++ actions/student/handleWithdrawAction.js
@@ -3,7 +3,9 @@
 import { auth } from "@/lib/auth";
 import dbConnect from "@/lib/dbConnect";
 import Application from "@/lib/models/application-model";
+import Internship from "@/lib/models/internship-model";
 
+import closeInternship from "@/lib/utils/closeInternship";
 export default async function handleWithdrawAction(internshipId) {
   try {
     await dbConnect();
@@ -13,6 +15,19 @@ export default async function handleWithdrawAction(internshipId) {
       return { error: "not logged in as student." };
     }
     const studentId = session.user.userId;
+
+    const internship = await Internship.findById(internshipId);
+
+    if (!internship) return { error: "Internship doesnt exist." };
+
+    const applicationDate = new Date(internship.application_date);
+
+    if (internship.isClosed || applicationDate < Date.now()) {
+      if (applicationDate < Date.now()) {
+        closeInternship(internshipId);
+      }
+      return { error: "Internship is already closed. Unable to withdraw." };
+    }
 
     const result = await Application.deleteOne({
       student_id: studentId,
